rename: InvalidUserInputException -> UserInvalidInputException
