+++ src/main/java/com/example/backend/global/exception/user/UserInvalidInputExceptionType.java
@@ -1,11 +1,11 @@
-package com.example.backend.global.exception;
+package com.example.backend.global.exception.user;
 
 import com.example.backend.global.exception.base.CustomExceptionType;
 
 /**
  * HttpStatus: BAD_REQUEST
  */
-public enum InvalidUserInputExceptionType implements CustomExceptionType {
+public enum UserInvalidInputExceptionType implements CustomExceptionType {
     ALREADY_EXISTS_EMAIL(-101,"'email'(body)이 이미 존재합니다."),
     ALREADY_EXISTS_NICKNAME(-102,"'nickname'(body)이 이미 존재합니다."),
     ALREADY_EXIST_EMAIL_AND_NICKNAME( -103,"'email'(body)과 'nickname'(body)이 이미 존재합니다."),
@@ -16,7 +16,7 @@ public enum InvalidUserInputExceptionType implements CustomExceptionType {
     private int errorCode;
     private String errorMessage;
 
-    InvalidUserInputExceptionType(int errorCode, String errorMessage) {
+    UserInvalidInputExceptionType(int errorCode, String errorMessage) {
         this.errorCode = errorCode;
         this.errorMessage = errorMessage;
     }
